♻️ Move extensions from View to Utils folders
