+++ Project/Challenge Alpha iOS/Challenge Alpha iOS/Scenes/PackageDetails/PackageDetailsView.swift
@@ -164,31 +164,3 @@ struct PackageDetailsView: View {
         .padding(.vertical)
     }
 }
-
-// MARK: - Extension to make applying AttributedString even easier
-extension Text {
-    init(_ string: String, configure: ((inout AttributedString) -> Void)) {
-        var attributedString = string.htmlToAttributedString()
-        configure(&attributedString)
-        self.init(attributedString)
-    }
-}
-
-extension StringProtocol {
-    func htmlToAttributedString() -> AttributedString {
-        guard let attString = try? AttributedString(
-            NSAttributedString(
-                data: .init(utf8),
-                options: [
-                    .documentType: NSAttributedString.DocumentType.html,
-                    .characterEncoding: String.Encoding.utf8.rawValue
-                ],
-                documentAttributes: nil
-            )
-        ) else {
-            return AttributedString(stringLiteral: "")
-        }
-        
-        return attString
-    }
-}
