+++ Project/Challenge Alpha iOS/Challenge Alpha iOS/UIComponents/Extensions/Text+Extension.swift
@@ -1,0 +1,18 @@
+//
+//  Text+Extension.swift
+//  Challenge Alpha iOS
+//
+//  Created by Yuri Strack on 18/01/23.
+//
+
+import Foundation
+import SwiftUI
+
+// MARK: - Configurable AttributedString inside Text View
+extension Text {
+    init(_ string: String, configure: ((inout AttributedString) -> Void)) {
+        var attributedString = string.htmlToAttributedString()
+        configure(&attributedString)
+        self.init(attributedString)
+    }
+}
