+++ Project/Challenge Alpha iOS/Challenge Alpha iOS/Utils/String+Extension.swift
@@ -30,3 +30,22 @@ extension String {
         return "\(monthString)/\(yearString)"
     }
 }
+
+extension StringProtocol {
+    func htmlToAttributedString() -> AttributedString {
+        guard let attString = try? AttributedString(
+            NSAttributedString(
+                data: .init(utf8),
+                options: [
+                    .documentType: NSAttributedString.DocumentType.html,
+                    .characterEncoding: String.Encoding.utf8.rawValue
+                ],
+                documentAttributes: nil
+            )
+        ) else {
+            return AttributedString(stringLiteral: "")
+        }
+        
+        return attString
+    }
+}
